Ensure chunk autonomy
Significant refactoring of code chunks was carried out with a view to
ensuring that each code chunk is autonomous once the setup chunk has been
run. This has become necessary so that code from chunks can be
independently parsed and evaluated elsewhere in the project e.g. in the
Shiny application.