Apply fix
An issue cropped up wherein reading some of the columns from the data
produced an error in the codebook. The problem appears to be in the
handling of labelled columns of class integer where missing values
recorded as NA had to be changed to NA_integer_ for the data to be read.