Fix bug: Improve extraction of label prefixes
When there are multi-response questions, the label prefixes are supposed to be removed so that the actual options are used in tabulation/plotting. Now, in the case of there the prefix-option separation is with a slash (/) that is not bounded by whitespace, the options could not be cleanly separated. The regular expressions used for this operation has now been fine tuned for a cleaner extraction of the relevant strings.

Relevant unit tests had to be added, and thus, the function doing this work, was moved into Namespace scope rather than within the table_multiopt function environment.